Add special overload for `useSelect` use case of select.
A special overload for handling `useSelect.

`useSelect` accepts either a string, or a callback function.
It is impossible to pass `typeof select`, and a callback in a type.
Instead, we overload the `select` and use it for both.

Technically this creates incorrect types when calling `select` directly
as it does not support a callback.